Don't use default is_staff column to check for core team members
Sometimes we want to allow some users to access the django admin
panel for some specific tasks (like the Minecraft server whitelist).
Django requires setting the is_staff column to True for that, but
we don't want thouse users to appear in the core team member list.
That's why we decided to add a new column for specifying if a user
belong to the core team or not.